Removed ability to read old (seems to me that it were Merc) areas. Area versions introduced. All ROM and SoG areas should be read successfully.
Rulesets reorganized. Now each object can have one ruleset
(instead of a number of separate rulesets for each keyword) and
exprs in `allow' and `deny' sets of rules can be written as
an expression of predicates with '&&', '||' and '!' operators.
lex+yacc parser for ruleset exprs.

extract_char now defaults to count extracted objs.

flag64_t is gone.
channels bits are kept separately from comm bits.
obj extra_flags are split into stat_flags and obj_flags.
obj_flags are static flags and are kept only in obj_index_data.
affected_by flags are split into affected_by, has_invis and has_detect
flags. races have has_invis and has_detect flags too.

fread_fword/fread_fstring now use flag_svalue.

backed out changes made between olc_rule.c 1.18 and 1.19
(automatic expl form creation in ruleed_create disabled).
automatic creation of expl forms using impl forms is back.

a few bugs in race OLC fixed (races were saved in unreadable format).


git-svn-id: file:///home/fjoe/old-repos/svn/husky/trunk/sog@1934 1738b366-ca78-dd11-ac90-000e8e03bdb1

+++ src/sog/obj_prog.c
@@ -1,5 +1,5 @@
 /*
- * $Id: obj_prog.c,v 1.73 1999-12-10 11:30:09 kostik Exp $
+ * $Id: obj_prog.c,v 1.74 1999-12-11 15:31:18 fjoe Exp $
  */
 
 /***************************************************************************
@@ -1608,7 +1608,7 @@ int wear_prog_fire_shield(OBJ_DATA *obj, CHAR_DATA *ch, const void *arg)
 	  af.duration = -2;
 	  af.level = LEVEL(ch);
 	  af.bitvector = 0;
-	  af.location = APPLY_RESIST_COLD;
+	  INT(af.location) = APPLY_RESIST_COLD;
 	  af.modifier = 50;
 	  affect_to_char(ch, &af);
 	}
@@ -1624,7 +1624,7 @@ int wear_prog_fire_shield(OBJ_DATA *obj, CHAR_DATA *ch, const void *arg)
 	  af.duration = -2;
 	  af.level = LEVEL(ch);
 	  af.bitvector = 0;
-	  af.location = APPLY_RESIST_FIRE;
+	  INT(af.location) = APPLY_RESIST_FIRE;
 	  af.modifier = 50;
 	  affect_to_char(ch, &af);
 	}
